+++ config/database.php
@@ -1,0 +1,10 @@
+<?php
+
+// Database
+define('GOTEO_DB_DRIVER',   'mysql');
+define('GOTEO_DB_HOST',     'localhost');
+define('GOTEO_DB_PORT',     3306);
+define('GOTEO_DB_CHARSET',  'UTF-8');
+define('GOTEO_DB_SCHEMA',   'db-schema');
+define('GOTEO_DB_USERNAME', 'db-username');
+define('GOTEO_DB_PASSWORD', 'db-password');
