+++ config/mail.php
@@ -1,0 +1,42 @@
+<?php
+
+/******************************************************
+PhpMailer constants
+*******************************************************/
+if (!defined('PHPMAILER_CLASS')) {
+    define ('PHPMAILER_CLASS', GOTEO_PATH . 'library' . DIRECTORY_SEPARATOR . 'phpmailer' . DIRECTORY_SEPARATOR . 'class.phpmailer.php');
+}
+if (!defined('PHPMAILER_LANGS')) {
+    define ('PHPMAILER_LANGS', GOTEO_PATH . 'library' . DIRECTORY_SEPARATOR . 'phpmailer' . DIRECTORY_SEPARATOR . 'language' . DIRECTORY_SEPARATOR);
+}
+if (!defined('PHPMAILER_SMTP')) {
+    define ('PHPMAILER_SMTP', GOTEO_PATH . 'library' . DIRECTORY_SEPARATOR . 'phpmailer' . DIRECTORY_SEPARATOR . 'class.smtp.php');
+}
+if (!defined('PHPMAILER_POP3')) {
+    define ('PHPMAILER_POP3', GOTEO_PATH . 'library' . DIRECTORY_SEPARATOR . 'phpmailer' . DIRECTORY_SEPARATOR . 'class.pop3.php');
+}
+
+//Mail management: ses (amazon), phpmailer (php library)
+define("MAIL_HANDLER", "phpmailer");
+
+// Mail
+define('GOTEO_MAIL_FROM', 'noreply@example.com');
+define('GOTEO_MAIL_NAME', 'example.com');
+define('GOTEO_MAIL_TYPE', 'smtp'); // mail, sendmail or smtp
+
+define('GOTEO_MAIL_SMTP_AUTH',     true);
+define('GOTEO_MAIL_SMTP_SECURE',   'ssl');
+define('GOTEO_MAIL_SMTP_HOST',     'smtp--host');
+define('GOTEO_MAIL_SMTP_PORT',     25); // Default smpt port
+define('GOTEO_MAIL_SMTP_USERNAME', 'smtp-usermail');
+define('GOTEO_MAIL_SMTP_PASSWORD', 'smtp-password');
+
+define('GOTEO_MAIL',         'info@example.com');
+define('GOTEO_CONTACT_MAIL', 'info@example.com');
+define('GOTEO_FAIL_MAIL',    'fail@example.com');
+define('GOTEO_LOG_MAIL',     'sitelog@example.com');
+
+//Quota de envio máximo para goteo en 24 horas
+define('GOTEO_MAIL_QUOTA', 50000);
+//Quota de envio máximo para newsletters para goteo en 24 horas
+define('GOTEO_MAIL_SENDER_QUOTA', round(GOTEO_MAIL_QUOTA * 0.8));
