+++ config/meta.php
@@ -1,0 +1,8 @@
+<?php
+
+// Metadata
+define('GOTEO_META_TITLE', '--meta-title--');
+define('GOTEO_META_DESCRIPTION', '--meta-description--');
+define('GOTEO_META_KEYWORDS', '--keywords--');
+define('GOTEO_META_AUTHOR', '--author--');
+define('GOTEO_META_COPYRIGHT', '--copyright--');
